+++ src/app/view/secure-pages/login/login.component.ts
@@ -7,6 +7,8 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { AuthService } from '../../../services/auth.service';
 import { RecaptchaModule, RecaptchaFormsModule } from 'ng-recaptcha';
 import { CategoryService } from "../../../services/category.service";
+import { ToastService } from "../../../services/toast.service";
+
 @Component({
   selector: 'app-login',
   standalone: true,
@@ -18,38 +20,48 @@ import { CategoryService } from "../../../services/category.service";
 })
 export class LoginComponent {
   fb = inject(FormBuilder);
-  router = inject(Router)
+  router = inject(Router);
   authService = inject(AuthService);
   categoryService = inject(CategoryService);
   loginForm!: FormGroup;
   passwordFieldType: string = 'password';
+  toastService = inject(ToastService);
 
+
+  
   ngOnInit(): void {
     this.loginForm = this.fb.group({
       email: ['', Validators.compose([Validators.required, Validators.email])],
       password: ['', Validators.required],
-      recaptcha: [null, Validators.required],
-      rememberMe: [false]
-    })
+      //recaptcha: [null, Validators.required],  
+      rememberMe: [false, Validators.requiredTrue]  // Making rememberMe checkbox mandatory
+    });
   }
 
   togglePasswordVisibility(): void {
     this.passwordFieldType = this.passwordFieldType === 'password' ? 'text' : 'password';
   }
-  resolved(captchaResponse:any) {
-    console.log(`Resolved captcha with response: ${captchaResponse}`);
-  }
+
+  // resolved(captchaResponse: any) {
+  //   console.log(`Resolved captcha with response: ${captchaResponse}`);
+  // }
+
   submit() {
-    console.log(this.loginForm.value);
+ 
+    
+    if (this.loginForm.invalid) {
+      this.toastService.show('Failed', 'Please fill in all required fields including the remember Me checkbox');
+      return;
+    }
     this.authService.loginService(this.loginForm.value)
       .subscribe({
         next: (res) => {
-          alert("Login Successfully");
-          console.log(res.data.roles);
-  
-          // Always clear sessionStorage to ensure no data remains
-          sessionStorage.clear();
-  
+        
+          this.toastService.show('Success', 'Login successfully!');
+         
+
+          sessionStorage.clear();  // Always clear sessionStorage
+
           if (this.loginForm.value.rememberMe) {
             localStorage.setItem("user_id", res.data._id);
             localStorage.setItem("token", res.token);
@@ -57,10 +69,9 @@ export class LoginComponent {
             localStorage.setItem("isBuyer", res.data.isBuyer);
             localStorage.setItem("isSeller", res.data.isSeller);
           } else {
-            // Clear localStorage if rememberMe is not checked
             localStorage.clear();
           }
-  
+
           if (res.data.isBuyer) {
             this.router.navigate(['/home']);
             this.categoryService.setHideCategory(false);
@@ -73,8 +84,9 @@ export class LoginComponent {
         },
         error: (err) => {
           console.log(err);
+         
         }
       });
   }
-  
 }
+
